[driver] Move modm::Terminal out of platform
This removes the dependency on boost for the hosted unittests.
