+++ src/modm/debug/module.lb
@@ -21,7 +21,7 @@ def prepare(module, options):
     target = options[":target"].identifier
     if target["platform"] == "hosted":
         # FIXME: Move logger/hosted/default_style.cpp into platform
-        module.depends(":platform:uart")
+        module.depends(":driver:terminal")
 
     module.depends(
         ":architecture",
