+++ src/modm/driver/io/terminal.lb
@@ -1,0 +1,23 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+#
+# Copyright (c) 2019, Niklas Hauser
+#
+# This file is part of the modm project.
+#
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+# -----------------------------------------------------------------------------
+
+def init(module):
+    module.name = ":driver:terminal"
+    module.description = "Terminal Output"
+
+def prepare(module, options):
+    module.depends(":io")
+    return True
+
+def build(env):
+    env.outbasepath = "modm/src/modm/driver/io/"
+    env.copy(".")
