+++ src/modm/platform/uart/hosted/module.lb
@@ -29,27 +29,16 @@ def prepare(module, options):
 
 
 def build(env):
-    device = env[":target"]
-    driver = device.get_driver("uart:hosted")
-
-    properties = device.properties
-    properties["target"] = target = device.identifier
-    properties["driver"] = driver
-
-    env.substitutions = properties
-    env.outbasepath = "modm/src/modm/platform/uart"
-
-    # FIXME: Ignore serial_interface, serial_port for windows
-    ignore = []
-    if env[":target"].identifier["family"] == "windows":
-        ignore.extend(["*/serial_interface*", "*/serial_port*"])
+    # FIXME: boost has several annoying warnings
+    env.collect(":build:ccflags.release", "-Wno-overloaded-virtual")
+    # FIXME: boost/type_traits/detail/config.hpp:85:69: warning: "__clang_major___WORKAROUND_GUARD" is not defined, evaluates to 0
+    env.collect(":build:cppdefines.release", "__clang_major___WORKAROUND_GUARD=4")
 
     env.collect(":build:library", "boost_system")
-    env.collect(":build:library",
-        "boost_thread" if env[":target"].identifier["family"] == "linux" else
-        "boost_thread-mt")
+    if env[":target"].identifier.family == "linux":
+        env.collect(":build:library", "boost_thread", "pthread")
+    else:
+        env.collect(":build:library", "boost_thread-mt")
 
-    if env[":target"].identifier["family"] == "linux":
-        env.collect(":build:library", "pthread")
-
-    env.copy(".", ignore=env.ignore_paths(*ignore))
+    env.outbasepath = "modm/src/modm/platform/uart"
+    env.copy(".")
