+++ tools/devices/hosted/windows.xml
@@ -9,6 +9,5 @@
     <driver name="core" type="hosted"/>
     <driver name="graphics" type="hosted"/>
     <driver name="gpio" type="hosted"/>
-    <driver name="uart" type="hosted"/>
   </device>
 </modm>
