+++ tools/scripts/synchronize_docs.py
@@ -61,7 +61,7 @@ def name(raw_name):
                    .replace("BLOCK-DEVICE-", "")\
                    .replace("BLOCK-", "")\
                    .replace("-SPI", "")
-    if result in ["DEVICE", "LIS3-TRANSPORT", "MEMORY-BUS",
+    if result in ["DEVICE", "LIS3-TRANSPORT", "MEMORY-BUS", "TERMINAL",
                   "MIRROR", "ADC-SAMPLER", "FAT", "HEAP", "--PYCACHE--"]:
         return None
     return result
